requirements.txt
psycopg2-binary==2.9.9
pyserial==3.5
python-dotenv==1.0.1
